failed attempts
form button was not responding to being clicked. attempted various combinations of javascript and jquery with no success. Attempted a simple alert command to test that button was even working and it would not respond with an alert. Removed check box, rebuilt with new class names. tested toggle command, seems to be working for now. will attempt to input a if else statement to populate fields.