form validation disable fields 2
Still encountering problems trying to get form fields to disable upon validation, getting closer to a solution,  trying different combinations of event handlers and playing around with applying various classes to if Else statement to trigger disable fields. Trying valiantly to arrive at a solution but this is quite a pickle.